+++ raft/src/main/java/org/apache/kafka/raft/EpochState.java
@@ -16,8 +16,6 @@
  */
 package org.apache.kafka.raft;
 
-import org.apache.kafka.raft.internals.ReplicaKey;
-
 import java.io.Closeable;
 import java.util.Optional;
