+++ raft/src/main/java/org/apache/kafka/raft/FollowerState.java
@@ -21,7 +21,6 @@ import org.apache.kafka.common.network.ListenerName;
 import org.apache.kafka.common.utils.LogContext;
 import org.apache.kafka.common.utils.Time;
 import org.apache.kafka.common.utils.Timer;
-import org.apache.kafka.raft.internals.ReplicaKey;
 import org.apache.kafka.snapshot.RawSnapshotWriter;
 
 import org.slf4j.Logger;
